+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * refactor: `DomainEventManagerInterface`
 * refactor: move interfaces to top level namespace
 * refactor: remove `DomainEventEmitterCollector`
+* feat: Now should fully support multiple entity managers.
 
 ## 1.2.5